e2e: node: cpumanager: move specific helpers
We're nearing the planned replacement of old cpumanager
tests with new ones, so move the remaining utilities
which we will keep using with the new codebase.

Trivial code movement, no intended behavioral changes.
